+++ TrelloWebApplication/Models/Badge.cs
@@ -7,14 +7,6 @@ namespace TrelloWebApplication.Models
 {
     public class Badge
     {
-        public string Votes { get; set; }
-        public string ViewingMemberVoted { get; set; }
-        public string Subscribed { get; set; }
-        public string Fogbugz { get; set; }
-        public string Due { get; set; }
-        public string Description { get; set; }
-        public string Comments { get; set; }
-        public string CheckItemsChecked { get; set; }
         public string CheckItems { get; set; }
         public string Attachments { get; set; }
     }
