+++ TrelloWebApplication/TrelloWebApplication.csproj
@@ -174,6 +174,7 @@
     <Compile Include="Models\Label.cs" />
     <Compile Include="Models\List.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Utiliti\PopolateModel.cs" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="Content\bootstrap-theme.css" />
